yolo-service/requirements.txt
flask==2.3.3
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
requests==2.31.0
torch>=1.8.0
torchvision>=0.9.0
